+++ scripts/download.js
@@ -34,7 +34,40 @@ async function download(name, year, filename, folder, options) {
         if (options?.fixTitle) {
             if (item.title.toUpperCase() === item.title) {
                 item.title = toTitleCase(item.title);
-            }          
+            }
+        }
+
+        // Clear unneeded properties
+        item.program = undefined;
+        item.donation_link = undefined;
+        item.guided_tours = undefined;
+        item.self_guided_tour_map = undefined;
+        item.contact_email = undefined;
+        if (item.event_type) {
+            item.event_type.id = undefined;
+            item.event_type.abbr = undefined;
+        }
+        item.slug = undefined;
+
+        if (item.description) {
+            if (!['.','!','?'].includes(item.description[item.description.length-1])) {
+                item.description = item.description + '.';
+                console.warn(`Added full stop to description of ${(options?.fixTitle) ? item.title : item.name }`);
+            }
+            if (item.description[0].toUpperCase() != item.description[0]) {
+                item.description = item.description.charAt(0).toUpperCase()
+                + item.description.slice(1);
+                console.warn(`Capitalized description of ${item.name}`);
+            }
+        }
+        if (item.print_description) {
+            if (!item.print_description.endsWith('.')) {
+                item.print_description += '.';
+            }
+            if (item.print_description[0].toUpperCase() != item.print_description[0]) {
+                item.print_description = item.print_description.charAt(0).toUpperCase()
+                + item.print_description.slice(1);
+            }
         }
         if (options?.fixOccurrence) {
             if (!item.occurrence_set) {
@@ -48,7 +81,14 @@ async function download(name, year, filename, folder, options) {
     json = json.filter((item) => !item.invalid);
 
     const f = `./src/assets/${folder}/${filename}.json`;
-    save(f, folder, JSON.stringify(json, undefined, 2));
+    const data = JSON.stringify(json, undefined, 2);
+    const changed = compare(f, folder, data);
+    if (changed) {
+        save(f, folder, data);
+    } else {
+        console.log(`No changes in ${folder} ${filename}`);
+    }
+    return changed;
 }
 
 function toTitleCase(str) {
@@ -84,6 +124,16 @@ function save(path, folder, data) {
     console.log(`Wrote "${otherPath}"`);
 }
 
+function compare(path, folder, data) {
+    const filename = basename(path);
+    if (!existsSync(path)) {
+        console.log(`${path} is missing.`);
+        return true;
+    }
+    const read = readFileSync(path, 'utf-8');
+    return read !== data;
+}
+
 function getUrl(name, year) {
     return `https://api.burningman.org/api/v1/${name}?year=${year}`;
 }
@@ -93,9 +143,11 @@ const years = process.argv.splice(2);
 console.log(years);
 for (const year of years) {
     console.log(`Downloading ${year}`);
-    await download('art', year, 'art', `ttitd-${year}`, { fixName: true });
-    await download('camp', year, 'camps', `ttitd-${year}`, { fixName: true });
-    await download('event', year, 'events', `ttitd-${year}`, { fixOccurrence: true, fixTitle: true, fixUid: true });
-    saveRevision(`ttitd-${year}`);
+    const artChanged = await download('art', year, 'art', `ttitd-${year}`, { fixName: true });
+    const campsChanged = await download('camp', year, 'camps', `ttitd-${year}`, { fixName: true });
+    const eventsChanged = await download('event', year, 'events', `ttitd-${year}`, { fixOccurrence: true, fixTitle: true, fixUid: true });
+    if (artChanged || campsChanged || eventsChanged) {
+        saveRevision(`ttitd-${year}`);
+    }
 }