+++ src/app/category/category.component.ts
@@ -10,7 +10,8 @@ import { IonicModule } from '@ionic/angular';
   imports: [IonicModule, NgFor]
 })
 export class CategoryComponent implements OnInit {
-
+  @Input() id = '';
+  @Input() allTitle: string = '';
   @Input() category: string = '';
   @Input() categories: string[] = [];
   @Output() categoryChange = new EventEmitter<string>();