+++ src/app/event/event.component.ts
@@ -1,14 +1,16 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { IonicModule } from '@ionic/angular';
 import { Event } from '../models';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-event',
   templateUrl: './event.component.html',
   styleUrls: ['./event.component.scss'],
   standalone: true,
-  imports: [IonicModule, RouterModule]
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [IonicModule, CommonModule, RouterModule]
 })
 export class EventComponent  implements OnInit {