+++ src/app/favorites.service.ts
@@ -3,6 +3,9 @@ import { Event, Favorites, OccurrenceSet } from './models';
 import { NotificationService } from './notification.service';
 import { Preferences } from '@capacitor/preferences';
 import { SettingsService } from './settings.service';
+import { DbService } from './db.service';
+import { getDayName, getOccurrenceTimeString, now } from './utils';
+import { MapPoint } from './map/map.component';
 
 enum DbId {
   favorites = 'favorites'
@@ -16,10 +19,11 @@ export class FavoritesService {
   private ready: Promise<void> | undefined;
   public changed = signal(1);
   private dataset: string = '';
+  private mapPoints: MapPoint[] = [];
 
   private favorites: Favorites = { art: [], events: [], camps: [], friends: [] };
 
-  constructor(private notificationService: NotificationService, private settingsService: SettingsService) {
+  constructor(private notificationService: NotificationService, private settingsService: SettingsService, private db: DbService) {
     this.init(this.settingsService.settings.dataset);
   }
 
@@ -27,6 +31,14 @@ export class FavoritesService {
   public async init(dataset: string) {
     this.dataset = dataset;
     this.ready = this.load();
+  }
+
+  public setMapPoints(mapPoints: MapPoint[]) {
+    this.mapPoints = mapPoints;
+  }
+
+  public getMapPoints(): MapPoint[] {
+    return this.mapPoints;
   }
 
   private noData(): Favorites {
@@ -104,6 +116,59 @@ export class FavoritesService {
     await this.saveFavorites();
   }
 
+  public async getEventList(ids: string[], historical: boolean): Promise<Event[]> {
+    const events = await this.db.getEventList(this.eventsFrom(ids));
+    // Group events and Set event time string to favorited event occurrence
+    const eventItems = await this.splitEvents(events, historical);
+    this.groupEvents(eventItems);
+    return eventItems;
+  }
+
+  private async splitEvents(events: Event[], historical: boolean): Promise<Event[]> {
+    const eventItems: Event[] = [];
+    const timeNow = now().getTime();
+    for (let event of events) {
+      for (let occurrence of event.occurrence_set) {
+        occurrence.star = await this.isFavEventOccurrence(event.uid, occurrence);
+        if (occurrence.star) {
+          const eventItem = structuredClone(event);
+          eventItem.occurrence_set = [structuredClone(occurrence)];
+
+          let start: Date = new Date(occurrence.start_time);
+          let end: Date = new Date(occurrence.end_time);
+
+          const isOld = (end.getTime() - timeNow < 0);
+          const isHappening = (start.getTime() < timeNow) && !isOld;
+          eventItem.occurrence_set[0].old = isOld;
+          eventItem.occurrence_set[0].happening = isHappening;
+          // console.log(eventItem.title);
+          // console.log(`Ends ${end} (${end.getTime()}), now=${now()} (${timeNow}), isHappening=${isHappening} isOld=${isOld}`);
+          eventItem.old = isOld;
+          eventItem.happening = isHappening;
+          const times = getOccurrenceTimeString(start, end, undefined);
+          eventItem.timeString = times ? times?.short : '';
+          eventItem.longTimeString = times ? times?.long : '';
+          if (!eventItem.old || historical) {
+            eventItems.push(eventItem);
+          }
+        }
+      }
+    }
+    eventItems.sort((a, b) => { return Date.parse(a.occurrence_set[0].start_time) - Date.parse(b.occurrence_set[0].start_time); });
+    return eventItems;
+  }
+
+  private groupEvents(events: Event[]) {
+    let group = '';
+    for (let event of events) {
+      const day = getDayName(event.occurrence_set[0].start_time);
+      if (day !== group) {
+        group = day;
+        event.group = group;
+      }
+    }
+  }
+
   private async saveFavorites() {
     const id = DbId.favorites;
     const value = this.favorites;
@@ -125,12 +190,9 @@ export class FavoritesService {
     }
   }
 
-
-
   private async load() {
     try {
       this.favorites = JSON.parse(await this.get(DbId.favorites, this.favorites));
-      console.log('loaded', this.favorites);
     } catch {
       this.favorites = this.noData();
     }