+++ changelog.md
@@ -1,4 +1,20 @@
 ## Changes
 
+### 1.2
+- Favorite filter added for events, art, camps, all
+- Add elipsis to event time listing to prevent overflow of text
+- Favorites are now grouped by the day they happen and are split based on occurrence
+- Fix for event showing the first occurrence not the favorited one
+- Display badge for "now" when an event is happening now
+- Favorite events that have passed do not show in the list
+- Fix to correctly choose todays day if its available in events
+- Search now has a "go" button
+- Map of Favorite Art, Events and Camps
+- Show events that are happening now
+- Better message when no events match
+- Favorites can be searched
+- Events that happened will disappear
+- Share app feature added
+
 ### 1.1 - August 2, 2023
 - Submitted to the App Store